move src/lib/platform.js to repo root fixes: Cannot find module './src/lib/platform' after packaging and publishing see: https://app.shortcut.com/particle/story/107005
