+++ settings.js
@@ -1,31 +1,9 @@
-/**
- ******************************************************************************
- * @file    settings.js
- * @author  David Middlecamp (david@particle.io)
- * @company Particle ( https://www.particle.io/ )
- * @source https://github.com/particle-iot/particle-cli
- * @version V1.0.0
- * @date    14-February-2014
- * @brief   Setting module
- ******************************************************************************
-  Copyright (c) 2016 Particle Industries, Inc.  All rights reserved.
-
-  This program is free software; you can redistribute it and/or
-  modify it under the terms of the GNU Lesser General Public
-  License as published by the Free Software Foundation, either
-  version 3 of the License, or (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-  Lesser General Public License for more details.
-
-  You should have received a copy of the GNU Lesser General Public
-  License along with this program; if not, see <http://www.gnu.org/licenses/>.
-  ******************************************************************************
- */
-
-const { PlatformId } = require('./src/lib/platform');
+///////////////////////////////////////////////////////////////////////////////
+// NOTE ///////////////////////////////////////////////////////////////////////
+// Due to how the cli is packaged and distributed, we cannot require files ////
+// from the `./src` directory /////////////////////////////////////////////////
+///////////////////////////////////////////////////////////////////////////////
+const { PlatformId } = require('./platform');
 
 const fs = require('fs');
 const path = require('path');
